+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [v0.7.0+obx.1.3.3]
+
+### Fixed
+
+- Fix pretrained weights not loaded properly in patchcore
+
+### Updated 
+
+- Use TimmFeatureExtractor instead of the deprecated FeatureExtractor
+
 ## [v0.7.0+obx.1.3.2]
 
 ### Fixed